refactor: TwoPainWithSlideBarLayout.vue のフォルダ階層を変更。
