+++ src/components/diagrams/editor/template/DiagramEditor.vue
@@ -65,7 +65,7 @@ import {
 } from "nuxt-property-decorator";
 
 import ResourceRightClickMenu from "./ResourceRightClickMenu.vue";
-import TwoPainWithSlideBarLayout from "@/components/TwoPainWithSlideBarLayout.vue";
+import TwoPainWithSlideBarLayout from "@/components/twopain/TwoPainWithSlideBarLayout.vue";
 import DiagramCanvas from "@/components/diagrams/editor/template/canvas/DiagramCanvas.vue";
 import ResourceParet from "@/components/diagrams/editor/template/paret/ResourceParet.vue";
 import ResourceEditDialog from "@/components/resource/ResourceEditDialog.vue";
